Sequelize logic moved to ./share folder, directde only for gateway
