+++ binance-smart-chain-rpc/src/main.js
@@ -23,7 +23,7 @@ module.exports.run = async () => {
     ethereum,
   }
 
-  if (process.env.MIGRATE_ON_BOOTSTRAP === 'true') await require('@/share/bin/sequelizeMigrate')()
+  if (process.env.MIGRATE_ON_BOOTSTRAP === 'true') await require('@/share/sequelize/bin/sequelizeMigrate')()
 
   console.log('[ethereum-rpc] Starting Express...')
   await require('@/constructors/express')(context)
