+++ gateway/src/Dockerfile
@@ -12,7 +12,6 @@ RUN apk add --no-cache \
 
 ADD ./gateway/src /app
 ADD ./share /app/share
-ADD ./share/models /app/constructors/sequelize/models
 
 RUN npm install
