--- gateway/src/bin/sequelizeSeed.js
@@ -1,15 +1,0 @@
-const cp = require('child_process')
-
-module.exports = async () => {
-  console.log('[Sequelize-Seed] Running seeds...')
-  try {
-    await cp.spawnSync('./node_modules/.bin/sequelize', ['db:seed:all'], {
-      env: process.env,
-      stdio: 'inherit',
-    })
-    console.log('[Sequelize-Migrate] Success.')
-  } catch (error) {
-    console.error(error)
-  }
-}
-
