+++ share/sequelize/bin/sequelizeMigrate.js
