+++ share/sequelize/config/postgres.js
