+++ share/sequelize/index.js
