+++ share/sequelize/seeders/20200920195019-EthereumContract.js
